+++ src/FluentDbTools/Implementations/FluentDbTools.Migration.Oracle/CustomProcessor/DefaultOracleCustomMigrationProcessor.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using FluentDbTools.Common.Abstractions;
 using FluentDbTools.Migration.Abstractions;
 using FluentDbTools.Migration.Abstractions.ExtendedExpressions;
+using FluentDbTools.Migration.Contracts.MigrationExpressions;
 using FluentMigrator.Model;
 using FluentMigrator.Runner.Processors.Oracle;
 using Microsoft.Extensions.Logging;
@@ -19,19 +20,19 @@ namespace FluentDbTools.Migration.Oracle.CustomProcessor
         /// <summary>
         /// IsNullable for "TIDSSTEMPEL" and "DIPS" collumn. Default is TRUE
         /// </summary>
-        [SuppressMessage("ReSharper", "IdentifierTypo")] 
-        public virtual bool IsTidsStempelAndDipsIdIdNullable { get; set; } = true;
+        [SuppressMessage("ReSharper", "IdentifierTypo")]
+        public bool IsTidsStempelAndDipsIdIdNullable { get; set; } = true;
 
         /// <summary>
         /// The SchemaName of CommonSchema (Default "DIPSCOREDB")
         /// </summary>
-        [SuppressMessage("ReSharper", "StringLiteralTypo")] 
-        public virtual string CommonSchemaName { get; set; } = "DIPSCOREDB";
+        [SuppressMessage("ReSharper", "StringLiteralTypo")]
+        public string CommonSchemaName { get; set; } = "DIPSCOREDB";
 
         /// <summary>
         /// The SchemaName of AuthCodeSchema (Default "CORE_CODE")
         /// </summary>
-        public virtual string AuthCodeSchemaName { get; set; } = "CORE_CODE";
+        public string AuthCodeSchemaName { get; set; } = "CORE_CODE";
 
         /// <summary>
         /// The Logger
@@ -46,7 +47,7 @@ namespace FluentDbTools.Migration.Oracle.CustomProcessor
         /// <summary>
         ///  Contains important Migration config values
         /// </summary>
-        protected IDbMigrationConfig MigrationConfig { get;}
+        protected IDbMigrationConfig MigrationConfig { get; }
 
         /// <summary>
         /// Is Only enabled if DbType is Oracle
@@ -75,18 +76,34 @@ namespace FluentDbTools.Migration.Oracle.CustomProcessor
         public virtual void MigrationMetadataChanged(IMigrationMetadata migrationMetadata, IExtendedMigrationProcessor extendedMigrationProcessor)
         {
             Processor = extendedMigrationProcessor;
-            if (!Processor.SchemaExists(CommonSchemaName) || 
+
+            if (!Enabled)
+            {
+                return;
+            }
+
+            if (!Processor.SchemaExists(CommonSchemaName) ||
                 !Processor.SchemaExists(AuthCodeSchemaName))
             {
                 return;
             }
 
-            var logonSql = SqlResources.LogonScriptSql
+            var logonSql = MigrationConfig
+                .PrepareSql(SqlResources.LogonScriptSql)
                 .ReplaceIgnoreCase("{CommonSchemaName}", CommonSchemaName)
-                .ReplaceIgnoreCase("{AuthCodeSchemaName}", AuthCodeSchemaName)
-                .ReplaceIgnoreCase("{User}", migrationMetadata.MigrationName.SubstringTo("."));
+                .ReplaceIgnoreCase("{AuthCodeSchemaName}", AuthCodeSchemaName);
 
             Execute(logonSql);
+
+            if (Processor.SchemaExists(MigrationConfig.Schema))
+            {
+                ProcessAfter(new CreateSchemaWithPrefixExpression
+                {
+                    SchemaName = MigrationConfig.Schema,
+                    SchemaPrefixId = MigrationConfig.GetSchemaPrefixId(),
+                    SchemaPrefixUniqueId = MigrationConfig.GetSchemaPrefixUniqueId()
+                });
+            }
         }
 
         /// <inheritdoc />
@@ -98,6 +115,11 @@ namespace FluentDbTools.Migration.Oracle.CustomProcessor
         /// <inheritdoc />
         public virtual void Process(IChangeLogTabledExpression expression)
         {
+            if (!Enabled)
+            {
+                return;
+            }
+
             if (expression.ChangeLogContext == null)
             {
                 return;
@@ -128,18 +150,23 @@ namespace FluentDbTools.Migration.Oracle.CustomProcessor
         /// <param name="expression"></param>
         public virtual void ProcessAfter(ICreateSchemaWithPrefixExpression expression)
         {
-            Logger.LogInformation($"ProcessAfter Schema:{expression.SchemaName} SchemaPrefix:{expression.SchemaPrefixId}");
+            if (!Enabled)
+            {
+                return;
+            }
+
             if (!Processor.SchemaExists(CommonSchemaName))
             {
                 return;
             }
 
-            var sqlAll = SqlResources.SchemaPrefixSql;
+            Logger.LogInformation($"ProcessAfter Schema:{expression.SchemaName} SchemaPrefix:{expression.SchemaPrefixId}");
 
-            sqlAll = sqlAll.Replace("{CommonSchemaName}", CommonSchemaName);
-            sqlAll = sqlAll.Replace("{SchemaName}", expression.SchemaName.ToUpper());
-            sqlAll = sqlAll.Replace("{SchemaPrefixUniqueId}", expression.SchemaPrefixUniqueId);
-            sqlAll = sqlAll.Replace("{SchemaPrefixId}", expression.SchemaPrefixId);
+            var sqlAll = MigrationConfig
+                .PrepareSql(SqlResources.SchemaPrefixSql)
+                .ReplaceIgnoreCase("{CommonSchemaName}", CommonSchemaName)
+                .ReplaceIgnoreCase("{AuthCodeSchemaName}", AuthCodeSchemaName);
+
 
             Execute(sqlAll);
         }
@@ -151,6 +178,16 @@ namespace FluentDbTools.Migration.Oracle.CustomProcessor
         /// <param name="expression"></param>
         public virtual void ProcessAfter(IDropSchemaWithPrefixExpression expression)
         {
+            if (!Enabled)
+            {
+                return;
+            }
+
+            if (!Processor.SchemaExists(CommonSchemaName))
+            {
+                return;
+            }
+
             Logger.LogInformation($"ProcessAfter Schema:{expression.SchemaName} SchemaPrefix:{expression.SchemaPrefixId}");
 
         }
@@ -158,39 +195,31 @@ namespace FluentDbTools.Migration.Oracle.CustomProcessor
         /// <inheritdoc />
         public virtual string GenerateSql(ICreateSchemaWithPrefixExpression expression)
         {
-            if (!Processor.SchemaExists(CommonSchemaName))
-            {
-                return null;
-            }
-
-            var sqlAll = SqlResources.CreateSchemaSql;
-
-            sqlAll = sqlAll.Replace("{CommonSchemaName}", CommonSchemaName);
-            sqlAll = sqlAll.Replace("{SchemaName}", expression.SchemaName);
-            sqlAll = sqlAll.Replace("{SchemaPrefixUniqueId}", expression.SchemaPrefixUniqueId);
-            sqlAll = sqlAll.Replace("{SchemaPrefix}", expression.SchemaPrefixId);
-            
-            return sqlAll;
-
+            return null;
         }
 
         /// <inheritdoc />
         public virtual IList<ColumnDefinition> GetDefaultColumns(string tableName = null)
         {
-            if (Processor.SchemaExists(CommonSchemaName))
+            if (!Enabled)
             {
-                return new List<ColumnDefinition>
-                {
-                    new ColumnDefinition { Name = "TIDSSTEMPEL", TableName = tableName, ColumnDescription = "Timestamp for", Type = DbType.Int32, IsNullable = IsTidsStempelAndDipsIdIdNullable},
-                    new ColumnDefinition { Name = "OPPRETTETAV", TableName = tableName, ColumnDescription = "Row created by", Type = DbType.AnsiString, Size = 20, IsNullable = true},
-                    new ColumnDefinition { Name = "OPPRETTETTID", TableName = tableName, ColumnDescription = "Created timestamp", Type = DbType.Date, IsNullable = true},
-                    new ColumnDefinition { Name = "SISTENDRETAV", TableName = tableName, ColumnDescription = "Last created by", Type = DbType.AnsiString, Size = 20, IsNullable = true},
-                    new ColumnDefinition { Name = "SISTENDRETTID", TableName = tableName, ColumnDescription = "Last created timestamp", Type = DbType.Date, IsNullable = true},
-                    new ColumnDefinition { Name = "DIPSID", TableName = tableName, ColumnDescription = "Row sequence", Type = DbType.Int32, IsNullable = IsTidsStempelAndDipsIdIdNullable},
-                };
+                return null;
             }
 
-            return new List<ColumnDefinition>();
+            if (!Processor.SchemaExists(CommonSchemaName))
+            {
+                return null;
+            }
+
+            return new List<ColumnDefinition>
+            {
+                new ColumnDefinition { Name = "TIDSSTEMPEL", TableName = tableName, ColumnDescription = "Timestamp for", Type = DbType.Int32, IsNullable = IsTidsStempelAndDipsIdIdNullable},
+                new ColumnDefinition { Name = "OPPRETTETAV", TableName = tableName, ColumnDescription = "Row created by", Type = DbType.AnsiString, Size = 20, IsNullable = true},
+                new ColumnDefinition { Name = "OPPRETTETTID", TableName = tableName, ColumnDescription = "Created timestamp", Type = DbType.Date, IsNullable = true},
+                new ColumnDefinition { Name = "SISTENDRETAV", TableName = tableName, ColumnDescription = "Last created by", Type = DbType.AnsiString, Size = 20, IsNullable = true},
+                new ColumnDefinition { Name = "SISTENDRETTID", TableName = tableName, ColumnDescription = "Last created timestamp", Type = DbType.Date, IsNullable = true},
+                new ColumnDefinition { Name = "DIPSID", TableName = tableName, ColumnDescription = "Row sequence", Type = DbType.Int32, IsNullable = IsTidsStempelAndDipsIdIdNullable},
+            };
 
         }
     }