+++ src/FluentDbTools/Tests/Test.FluentDbTools.Migration/Resources/_ResourceHelper.cs
@@ -1,0 +1,8 @@
+﻿namespace Test.FluentDbTools.Migration.Resources
+{
+    // ReSharper disable once InconsistentNaming
+    internal static class _ResourceHelper
+    {
+        internal static readonly string Location = typeof(_ResourceHelper).Namespace;
+    }
+}